\name{codepoint-package}
\alias{codepoint-package}
\alias{codepoint}
\docType{package}
\title{
Stuff to make life easier when using OS Code-Point(R) Open data
}
\description{
Stuff to make life easier when using OS Code-Point(R) Open data
}
\details{
~~ An overview of how to use the package, including the most important functions ~~

getCodePoint
}
\author{
Author: Wei Liang Yap

Maintainer: Wei Liang Yap <yapweiliang@yahoo.co.uk>
}
\references{
https://www.ordnancesurvey.co.uk/business-and-government/products/code-point-open.html
}
\keyword{ package }
\seealso{
}
\examples{
CodePoint <- getCodePoint( "H:/DATASETS/OS/CodePoint/tmp",rds.output.folder="H:/DATASETS/OS/CodePoint" )
}
